cron-job/requirements.txt
Flask==1.0.2
google-api-python-client==1.7.4
google-cloud-datastore==1.7.1
google-cloud-bigquery==1.7.0
PyMySQL==0.9.2